+++ src/App.js
@@ -1,31 +1,12 @@
 import React from 'react';
-const todoList = [
-  {
-    id: 1,
-    title: "Complete assignment"
-  },
-  {
-    id: 2,
-    title: "Create branch"
-  },
-  {
-    id: 3,
-    title: "Push to GitHub"
-  },
-  {
-    id: 4,
-    title: "Create pull request"
-  }
-];
+import AddTodoForm from './AddTodoForm';
+import TodoList from './TodoList';
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function(toDoListElement){
-          return <li key={toDoListElement.id}>{toDoListElement.title}</li>;
-        })}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
 
   );
